requirements_versions.txt
setuptools==69.5.1  # temp fix for compatibility with some old packages
GitPython==3.1.32
Pillow==9.5.0
accelerate==0.21.0
blendmodes==2022
clean-fid==0.1.35
diskcache==5.6.3
einops==0.4.1
facexlib==0.3.0
fastapi==0.94.0
gradio==3.41.2
httpcore==0.15
inflection==0.5.1
jsonmerge==1.8.0
kornia==0.6.7
lark==1.1.2
numpy==1.26.2
omegaconf==2.2.3
open-clip-torch==2.20.0
piexif==1.1.3
protobuf==3.20.0
psutil==5.9.5
pytorch_lightning==1.9.4
resize-right==0.0.2
safetensors==0.4.2
scikit-image==0.21.0
spandrel==0.3.4
spandrel-extra-arches==0.1.1
tomesd==0.1.3
torch
torchdiffeq==0.2.3
torchsde==0.2.6
transformers==4.36.2
httpx==0.24.1
pillow-avif-plugin==1.4.3
albumentations==1.4.3
pydantic==1.10.15
diffusers==0.24.0
cachetools==5.3.2
colorama==0.4.6
torchvision
decord
tqdm
opencv-python
compel
easydict
rotary_embedding_torch
imageio[ffmpeg]
httpx[socks]
